infinite mutation station
take the brakes off

+++ crawl-ref/source/mutation.cc
@@ -1622,14 +1622,17 @@ bool mutate(mutation_type which_mutation, const string &reason, bool failMsg,
         }
     }
 
+    // total muts / 15 = chance to remove a mutation instead of add
+    // Don't do this 1 time in 3 or if a forced mutation
     if (mutclass == MUTCLASS_NORMAL
         && (which_mutation == RANDOM_MUTATION
             || which_mutation == RANDOM_XOM_MUTATION)
-        && x_chance_in_y(you.how_mutated(false, true), 15))
+        && x_chance_in_y(you.how_mutated(false, true), 15)
+        && !one_chance_in(3) && !force_mutation)
     {
-        // God gifts override mutation loss due to being heavily
-        // mutated.
-        if (!one_chance_in(3) && !god_gift && !force_mutation)
+        // If a god gift, we won't cause mutation loss, but we 
+        // also won't add a new random mutation
+        if (god_gift)
             return false;
         else
             return delete_mutation(RANDOM_MUTATION, reason, failMsg,
